+++ docs/make.jl
@@ -1,17 +1,13 @@
 using Documenter, Query
 
-# Install dependencies and precompile everything
-Pkg.add("DataFrames")
-using DataFrames
-
 makedocs(
 	modules = [Query],
-	format = :html,
 	sitename = "Query.jl",
 	pages = [
 		"Introduction" => "index.md",
 		"Getting Started" => "gettingstarted.md",
-		"Query Commands" => "querycommands.md",
+		"Standalone Query Commands" => "standalonequerycommands.md",
+		"LINQ Style Query Commands" => "linqquerycommands.md",		
 		"Data Sources" => "sources.md",
 		"Data Sinks" => "sinks.md",
 		"Experimental Features" => "experimental.md",
@@ -19,9 +15,5 @@ makedocs(
 )
 
 deploydocs(
-    deps = nothing,
-    make = nothing,
-    target = "build",
-    repo = "github.com/queryverse/Query.jl.git",
-    julia = "1.0"
+    repo = "github.com/queryverse/Query.jl.git"
 )